Move badge into locale/_components folder
