+++ portal/app/[locale]/_components/badge.tsx
